fix: use numeric month format to avoid Hebrew/Arabic bidi scrambling in dates

+++ app/trips/[id]/page.tsx
@@ -68,9 +68,9 @@ export default function TripDetailPage({ params }: { params: Promise<{ id: strin
         <div className="absolute bottom-0 left-0 right-0 p-4">
           <p className="text-white font-bold text-lg leading-tight drop-shadow">{trip.name}</p>
           <p className="text-white/80 text-xs mt-0.5 drop-shadow" dir="ltr">
-            {new Date(trip.startDate).toLocaleDateString(dateLocale, { day: 'numeric', month: 'short' })}
+            {new Date(trip.startDate).toLocaleDateString(dateLocale, { day: 'numeric', month: 'numeric' })}
             {' – '}
-            {new Date(trip.endDate).toLocaleDateString(dateLocale, { day: 'numeric', month: 'short', year: 'numeric' })}
+            {new Date(trip.endDate).toLocaleDateString(dateLocale, { day: 'numeric', month: 'numeric', year: 'numeric' })}
             {' · '}
             {nights} {t(nights === 1 ? 'night' : 'nights')}
           </p>
